IGNITE-17083 Main branch merged.
