+++ modules/storage-page-memory/src/main/java/org/apache/ignite/internal/storage/pagememory/mv/AbstractPageMemoryMvPartitionStorage.java
@@ -673,10 +673,8 @@ public abstract class AbstractPageMemoryMvPartitionStorage implements MvPartitio
 
     @Override
     public @Nullable RowId closestRowId(RowId lowerBound) throws StorageException {
-        try {
-            try (Cursor<VersionChain> cursor = versionChainTree.find(new VersionChainKey(lowerBound), null)) {
-                return cursor.hasNext() ? cursor.next().rowId() : null;
-            }
+        try (Cursor<VersionChain> cursor = versionChainTree.find(new VersionChainKey(lowerBound), null)) {
+            return cursor.hasNext() ? cursor.next().rowId() : null;
         } catch (Exception e) {
             throw new StorageException("Error occurred while trying to read a row id", e);
         }
