+++ modules/table/src/main/java/org/apache/ignite/internal/table/distributed/raft/snapshot/message/SnapshotMetaResponse.java
@@ -15,13 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.table.message;
+package org.apache.ignite.internal.table.distributed.raft.snapshot.message;
 
+import org.apache.ignite.internal.table.distributed.TableMessageGroup;
+import org.apache.ignite.network.NetworkMessage;
 import org.apache.ignite.network.annotations.Transferable;
+import org.apache.ignite.raft.jraft.entity.RaftOutter.SnapshotMeta;
 
 /**
- * Snapshot partition data request message.
+ * Snapshot meta response message.
  */
-@Transferable(TableMessageGroup.SNAPSHOT_MV_DATA_REQUEST)
-public interface SnapshotMvDataRequest extends SnapshotRequestMessage {
+@Transferable(TableMessageGroup.SNAPSHOT_META_RESPONSE)
+public interface SnapshotMetaResponse extends NetworkMessage {
+    /** Snapshot meta. */
+    SnapshotMeta meta();
 }
