+++ modules/table/src/main/java/org/apache/ignite/internal/table/distributed/raft/snapshot/message/SnapshotMvDataResponse.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.table.message;
+package org.apache.ignite.internal.table.distributed.raft.snapshot.message;
 
 import java.nio.ByteBuffer;
 import java.util.Collection;
@@ -23,6 +23,7 @@ import java.util.List;
 import java.util.UUID;
 import org.apache.ignite.hlc.HybridTimestamp;
 import org.apache.ignite.internal.schema.BinaryRow;
+import org.apache.ignite.internal.table.distributed.TableMessageGroup;
 import org.apache.ignite.network.NetworkMessage;
 import org.apache.ignite.network.annotations.Marshallable;
 import org.apache.ignite.network.annotations.Transferable;
@@ -48,7 +49,7 @@ public interface SnapshotMvDataResponse extends NetworkMessage {
         /** Individual row id. */
         UUID rowId();
 
-        /** List of {@link BinaryRow}s for a given {@link #rowId()} */
+        /** List of {@link BinaryRow}s for a given {@link #rowId()}ю */
         @Marshallable
         List<ByteBuffer> rowVersions();
